+++ bee-common/bee-packable/src/packable/vec.rs
@@ -18,7 +18,7 @@ impl<T: Packable> Packable for Vec<T> {
 
     fn pack<P: Packer>(&self, packer: &mut P) -> Result<(), PackError<Self::PackError, P::Error>> {
         // The length of any dynamically-sized sequence must be prefixed.
-        self.len().pack(packer).map_err(PackError::infallible)?;
+        (self.len() as u64).pack(packer).map_err(PackError::infallible)?;
 
         for item in self.iter() {
             item.pack(packer)?;
@@ -28,14 +28,14 @@ impl<T: Packable> Packable for Vec<T> {
     }
 
     fn packed_len(&self) -> usize {
-        0usize.packed_len() + self.iter().map(T::packed_len).sum::<usize>()
+        0u64.packed_len() + self.iter().map(T::packed_len).sum::<usize>()
     }
 
     fn unpack<U: Unpacker>(unpacker: &mut U) -> Result<Self, UnpackError<Self::UnpackError, U::Error>> {
         // The length of any dynamically-sized sequence must be prefixed.
-        let len = usize::unpack(unpacker).map_err(UnpackError::infallible)?;
+        let len = u64::unpack(unpacker).map_err(UnpackError::infallible)?;
 
-        let mut vec = Self::with_capacity(len);
+        let mut vec = Self::with_capacity(len as usize);
 
         for _ in 0..len {
             let item = T::unpack(unpacker)?;
